ITEP-21376 Adding secure headers for App Orch (#56)
Both App Service Proxy and VNC applications need this new
`secure-headers-app-orch` middleware. So rather than repeating it in
both places, it is better to centralize it here along side the main
secure-headers.

You will see the `secure-headers-app-orch` has a CSP not as permissive
as the main secure headers, as these extra src's are not needed.

Please note the Cross-Origin-Embedder-Policy is set to `unsafe-none` as
Keycloak does not yet return a CORP header. This is planned for EMF 3.1